requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
pydantic-settings
python-dotenv
openai
numpy
apscheduler
requests
slowapi
jsonschema
pytest
